student halv vejes

+++ src/main/java/school/system/student_attendance/repositories/StudentsRepo.java
@@ -7,5 +7,5 @@ import school.system.student_attendance.models.Students;
 
 @Qualifier("StudentsRepo")
 @Repository
-public interface StudentsRepo extends CrudRepository<Students, Long> {
+public interface StudentsRepo extends CrudRepository<Students, Integer> {
 }
